+++ packages/core/src/config/defaults.ts
@@ -1,22 +1,18 @@
-import { HealthCheck, TerminusOptions } from '@godaddy/terminus';
+import { HealthCheckMap, TerminusOptions } from '@godaddy/terminus';
 
-const second = 1000;
+const second_to_millisecond = 1000;
 
-export const defaultHealthCheck = (isProduction: boolean, healthcheckOptions?: Partial<TerminusOptions['healthChecks']>) => ({
+export const defaultHealthCheck = (
+  { isProduction, ...rest }: HealthCheckMap & { isProduction?: boolean; }
+) => ({
   verbatim: !isProduction,
   __unsafeExposeStackTraces: !isProduction,
-  '/healthcheck': ((input) => {
-    const { state } = input;
-    if (state.isShuttingDown) {
-      throw new Error('Server is shutting down. See ya!');
-    }
-  }) as HealthCheck,
-  ...healthcheckOptions,
+  ...rest,
 });
 
 export const defaultTerminusOptions: TerminusOptions = ({
   // We have ~20 seconds before receiving SIGKILL
-  timeout: 20 * second,
+  timeout: 20 * second_to_millisecond,
   signal: 'SIGTERM',
   useExit0: true,
 });